requirements.txt
Flask
htmx.org
eccodes
matplotlib

pyyaml
pygrib
